feat: add various dashboard components including SummaryCard, ExpensePieChart, CashFlowChart, BudgetStatus, and RecentTransactions; update NotificationPanel and Layout for improved functionality
